requirements.txt
streamlit
tensorflow
numpy
opencv-python
matplotlib
Pillow